roundtable_dockerize/hashtag_api/hashtag/requirements.txt
click
h11
pydantic
starlette
fastapi
PyJWT
passlib
uvicorn
python-dotenv
transformers
keyphrase_vectorizers
keybert